Add way to check if entity is considered as a fake player or not.
This is done by adding a new Event which can be used to mark Entities as FakePlayers but also asking each platform if they consider the Entity as a FakePlayer by default.

Forge: Check to see if entity is instanceof their built-in FakePlayer class.

Fabric: Similar to Forge but using reflection to see if Technici4n's FakePlayer API exists.
If the API does exist, register an Event listener which marks Entities [that are instances of his FakePlayer class] as FakePlayers.

See
FabricEntityHooks.setupTechnici4nFakePlayerAPI
FAPI-FakePlayer-API-PR: https://github.com/FabricMC/fabric/pull/3005